LocalFilesystem throws errors in a browser
