+++ packages/sqrl/src/simple/runSqrlTest.ts
@@ -6,7 +6,7 @@
 import { SimpleManipulator } from "./SimpleManipulator";
 import { createSimpleContext } from "../api/ctx";
 import { SqrlTest } from "../testing/SqrlTest";
-import { LocalFilesystem, Filesystem } from "../api/filesystem";
+import {  Filesystem } from "../api/filesystem";
 import { buildTestInstance } from "../testing/runSqrlTest";
 import { Instance, Execution } from "../api/execute";
 import * as path from "path";
@@ -14,6 +14,7 @@ import { Logger } from "../api/log";
 import { Config } from "../api/config";
 import { FunctionCostData } from "../function/Instance";
 import { invariant } from "sqrl-common";
+import { LocalFilesystem } from "../node/LocalFilesystem";
 
 export async function runSqrlTest(
   sqrl: string,
